Unify rate limiter implementation
